Re-organise test code #303
